+++ unit_test/tests/Node.cpp
@@ -1,8 +1,8 @@
-#include "../utilities/device_utilities.h"
-#include "../utilities/system_testings.h"
-#include "../utilities/test_utilities.h"
+#include "../test_utilities/device_utilities.h"
+#include "../test_utilities/system_testings.h"
+#include "../test_utilities/test_utilities.h"
 
-#include "../utilities/pch.h"
+#include "../test_utilities/pch.h"
 
 namespace {
 	class NodeTest : public ::testing::Test {
